Add primitive endorsement support
This lets peers run a custom validation code before transactions are
executed. This custom code is injected via Go plugins, which
unfortunately imposes a few practical limitations: most prominently,
plugins must be compiled with the same version of Go as burrow itself.
Using plugins prevents us from compiling Go as a static binary.
Also, unfortunately, not all platforms support Go plugins: for example,
Go on OpenBSD has no support for this feature.

Nevertheless, users may specify endorsers in burrow.toml by appending
to the list [[Execution.Endorsers]]. Each endorser definition consists
of a Name and a Library. The Name is of course an arbitrary string, but
the Library points to the compiled Go plugin (a .so file).

Every endorser implements the Endorser interface defined in the
execution module. It only has a single function, Endorse, which gets the
entire transaction envelope (and thus the payload) and returns a boolean
(and/or optionally an error). As expected, retuning with false causes
the transaction to be aborted. Errors also imply aborts.

This endorsement verification occurs in burrow's execution module, along
with other validation code.

There may be better ways to do this. This commit just implements some
basic functionality.

Signed-off-by: bzp99 <bertalan.peter@bp99.eu>